Move column addition tests to `BoardTest`
